+++ packages/extension-koni-ui/src/Popup/Confirmations/variants/Transaction/variants/ClaimReward.tsx
@@ -2,9 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { RequestStakeClaimReward } from '@subwallet/extension-base/background/KoniTypes';
-import CommonTransactionInfo from '@subwallet/extension-koni-ui/components/Confirmation/CommonTransactionInfo';
-import MetaInfo from '@subwallet/extension-koni-ui/components/MetaInfo';
-import useGetNativeTokenBasicInfo from '@subwallet/extension-koni-ui/hooks/common/useGetNativeTokenBasicInfo';
+import { CommonTransactionInfo, MetaInfo } from '@subwallet/extension-koni-ui/components';
+import { useGetNativeTokenBasicInfo } from '@subwallet/extension-koni-ui/hooks';
 import CN from 'classnames';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
