+++ packages/extension-koni-ui/src/components/History/HistoryItem.tsx
@@ -2,8 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { TransactionDirection } from '@subwallet/extension-base/background/KoniTypes';
-import { TransactionHistoryDisplayItem } from '@subwallet/extension-koni-ui/Popup/Home/History';
-import { ThemeProps } from '@subwallet/extension-koni-ui/types';
+import { ThemeProps, TransactionHistoryDisplayItem } from '@subwallet/extension-koni-ui/types';
 import { Icon, Logo, Number, Web3Block } from '@subwallet/react-ui';
 import CN from 'classnames';
 import { CaretRight } from 'phosphor-react';
