+++ packages/extension-koni-ui/src/components/MetaInfo/index.ts
@@ -1,0 +1,8 @@
+// Copyright 2019-2022 @subwallet/extension-koni-ui authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import { InfoItemBase as _InfoItemBase } from './parts';
+
+export type InfoItemBase = _InfoItemBase;
+
+export { default as MetaInfo } from './MetaInfo';
