+++ packages/extension-koni-ui/src/components/MetaInfo/parts/index.tsx
@@ -1,6 +1,7 @@
 // Copyright 2019-2022 @subwallet/extension-koni-ui authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+export { default as AccountGroupItem } from './AccountGroupItem';
 export { default as AccountItem } from './AccountItem';
 export { default as ChainItem } from './ChainItem';
 export { default as DataItem } from './DataItem';
@@ -10,3 +11,5 @@ export { default as NumberItem } from './NumberItem';
 export { default as StatusItem } from './StatusItem';
 export { default as TotalItem } from './TotalItem';
 export { default as TransferItem } from './TransferItem';
+
+export type { InfoItemBase } from './types';
