+++ packages/extension-koni-ui/src/components/index.ts
@@ -6,7 +6,6 @@ export { default as ChainItemFooter } from './ChainItemFooter';
 export { default as EmptyList } from './EmptyList';
 export { default as GeneralEmptyList } from './GeneralEmptyList';
 export { default as LoadingScreen } from './LoadingScreen';
-export { default as MetaInfo } from './MetaInfo';
 export { default as NetworkToggleItem } from './NetworkToggleItem';
 export { default as RadioGroup } from './Field/RadioGroup';
 export { default as ScreenTab } from './ScreenTab';
@@ -22,6 +21,7 @@ export * from './History';
 export * from './Icon';
 export * from './Layout';
 export * from './Logo';
+export * from './MetaInfo';
 export * from './Modal';
 export * from './Qr';
 export * from './Setting';
