+++ packages/extension-koni-ui/src/types/index.ts
@@ -96,6 +96,7 @@ export interface SigData {
 export * from './account';
 export * from './confirmation';
 export * from './form';
+export * from './history';
 export * from './navigation';
 export * from './network';
 export * from './staking';
