+++ packages/extension-koni-ui/src/util/transaction/index.ts
@@ -2,3 +2,4 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export * from './stake';
+export * from './detectType';
